# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RoomReview",
    "version": "1.0.0",
    "description": "API documentation for RoomReview"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/properties/{id}": {
      "get": {
        "summary": "Get property by ID",
        "tags": [
          "Properties"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Property details"
          }
        }
      },
      "put": {
        "summary": "Update property by ID",
        "tags": [
          "Properties"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Property updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete property by ID",
        "tags": [
          "Properties"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Property deleted successfully"
          }
        }
      }
    },
...
  },
...
}